backoffice/requirements.txt
absl-py==1.0.0
aiomysql==0.0.22
aiosqlite==0.17.0
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cycler==0.11.0
Deprecated==1.2.13
ecdsa==0.17.0
fastapi==0.70.1
fonttools==4.28.5
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
httpcore==0.14.4
httpx==0.21.3
idna==3.3
importlib-metadata==4.10.0
importlib-resources==5.4.0
iniconfig==1.1.1
kiwisolver==1.3.2
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
passlib==1.7.4
Pillow==9.0.0
pluggy==1.0.0
protobuf==3.19.3
psycopg2-binary==2.9.3
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
PyMySQL==0.9.3
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.1.0
rfc3986==1.5.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
sqlalchemy-schemadisplay==1.3
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.16.0
wrapt==1.13.3
zipp==3.7.0
